+++ tests/TestCase/Utility/ThumbCreatorTest.php
@@ -75,7 +75,7 @@ class ThumbCreatorTest extends TestCase
     /**
      * Test for `__construct()` method, passing a no existing file
      * @expectedException Cake\Network\Exception\InternalErrorException
-     * @expectedExceptionMessage File `webroot/img/noExistingFile.gif` not readable
+     * @expectedExceptionMessage File `tests/test_app/webroot/img/noExistingFile.gif` not readable
      * @test
      */
     public function testConstructNoExistingFile()
@@ -86,7 +86,7 @@ class ThumbCreatorTest extends TestCase
     /**
      * Test for `__construct()` method, passing a no existing file from plugin
      * @expectedException Cake\Network\Exception\InternalErrorException
-     * @expectedExceptionMessage File `Plugin/TestPlugin/webroot/img/noExistingFile.gif` not readable
+     * @expectedExceptionMessage File `tests/test_app/Plugin/TestPlugin/webroot/img/noExistingFile.gif` not readable
      * @test
      */
     public function testConstructNoExistingFileFromPlugin()